moved interpreting parsers to separate module called interpreted_mathml.rs, added support for parsing simple equations using the interpreting parser
