+++ skema/skema-rs/mathml/src/parsers.rs
@@ -1,3 +1,4 @@
 pub mod first_order_ode;
 pub mod generic_mathml;
+pub mod interpreted_mathml;
 pub mod math_expression_tree;
